Create blender manifest and build script
