+++ blender/extension/blender_manifest.toml
@@ -1,0 +1,93 @@
+schema_version = "1.0.0"
+
+id         = "protein_runway"
+version    = "0.0.1"
+name       = "Protein Runway"
+tagline    = "Render proteins with their trajectories, segment and extract normal modes."
+maintainer = "TODO <email@example.com>"
+
+type = "add-on"
+
+# # Optional: link to documentation, support, source files, etc
+# website = "https://extensions.blender.org/add-ons/my-example-package/"
+
+# # Optional: tag list defined by Blender and server, see:
+# # https://docs.blender.org/manual/en/dev/advanced/extensions/tags.html
+# tags = ["Animation", "Sequencer"]
+
+blender_version_min = "4.2.0"
+# # Optional: Blender version that the extension does not support, earlier versions are supported.
+# # This can be omitted and defined later on the extensions platform if an issue is found.
+# blender_version_max = "5.1.0"
+
+# https://docs.blender.org/manual/en/dev/advanced/extensions/licenses.html
+license = [
+  "SPDX:GPL-3.0-or-later",
+]
+# # Optional: required by some licenses.
+# copyright = [
+#   "2002-2024 Developer Name",
+#   "1998 Company Name",
+# ]
+
+platforms = ["windows-x64", "macos-arm64", "linux-x64", "macos-x64"]
+# Other supported platforms: "windows-arm64"
+
+# https://docs.blender.org/manual/en/dev/advanced/extensions/python_wheels.html
+#
+# All packages are dependencies of MDAnalysis
+#
+wheels = [
+  "./wheels/GridDataFormats-1.0.2-py3-none-any.whl",
+  "./wheels/MDAnalysis-2.7.0-cp312-cp312-manylinux_2_17_x86_64.manylinux2014_x86_64.whl",
+  "./wheels/contourpy-1.3.1-cp312-cp312-manylinux_2_17_x86_64.manylinux2014_x86_64.whl",
+  "./wheels/cycler-0.12.1-py3-none-any.whl",
+  "./wheels/fasteners-0.19-py3-none-any.whl",
+  "./wheels/fonttools-4.54.1-cp312-cp312-manylinux_2_17_x86_64.manylinux2014_x86_64.whl",
+  "./wheels/joblib-1.4.2-py3-none-any.whl",
+  "./wheels/kiwisolver-1.4.7-cp312-cp312-manylinux_2_17_x86_64.manylinux2014_x86_64.whl",
+  "./wheels/matplotlib-3.9.2-cp312-cp312-manylinux_2_17_x86_64.manylinux2014_x86_64.whl",
+  "./wheels/mda_xdrlib-0.2.0-py3-none-any.whl",
+  "./wheels/mmtf_python-1.1.3-py2.py3-none-any.whl",
+  "./wheels/mrcfile-1.5.3-py2.py3-none-any.whl",
+  "./wheels/msgpack-1.1.0-cp312-cp312-manylinux_2_17_x86_64.manylinux2014_x86_64.whl",
+  "./wheels/numpy-1.26.4-cp312-cp312-manylinux_2_17_x86_64.manylinux2014_x86_64.whl",
+  "./wheels/packaging-24.2-py3-none-any.whl",
+  "./wheels/pillow-11.0.0-cp312-cp312-manylinux_2_28_x86_64.whl",
+  "./wheels/pyparsing-3.2.0-py3-none-any.whl",
+  "./wheels/python_dateutil-2.9.0.post0-py2.py3-none-any.whl",
+  "./wheels/scipy-1.14.1-cp312-cp312-manylinux_2_17_x86_64.manylinux2014_x86_64.whl",
+  "./wheels/six-1.16.0-py2.py3-none-any.whl",
+  "./wheels/threadpoolctl-3.5.0-py3-none-any.whl",
+  "./wheels/tqdm-4.67.0-py3-none-any.whl",
+]
+
+# # Optional: add-ons can list which resources they will require:
+# # * files (for access of any filesystem operations)
+# # * network (for internet access)
+# # * clipboard (to read and/or write the system clipboard)
+# # * camera (to capture photos and videos)
+# # * microphone (to capture audio)
+# #
+# # If using network, remember to also check `bpy.app.online_access`
+# # https://docs.blender.org/manual/en/dev/advanced/extensions/addons.html#internet-access
+# #
+# # For each permission it is important to also specify the reason why it is required.
+# # Keep this a single short sentence without a period (.) at the end.
+# # For longer explanations use the documentation or detail page.
+#
+# [permissions]
+# network = "Need to sync motion-capture data to server"
+# files = "Import/export FBX from/to disk"
+# clipboard = "Copy and paste bone transforms"
+
+# # Optional: advanced build settings.
+# # https://docs.blender.org/manual/en/dev/advanced/extensions/command_line_arguments.html#command-line-args-extension-build
+# [build]
+# # These are the default build excluded patterns.
+# # You only need to edit them if you want different options.
+# paths_exclude_pattern = [
+#   "__pycache__/",
+#   "/.git/",
+#   "/*.zip",
+# ]
